+++ src/driveType/PeanutChassis.cpp
@@ -233,16 +233,16 @@ void PeanutChassis::driveStraight(double speed)
 
 /**
  * This will return the encoder ticks for the left and right motors. It will
- * return as a double pointer, but should be used an array. The first index [0]
+ * return as an integer pointer, but should be used an array. The first index [0]
  * is the left encoder tick count, and the second index [1] is the right
  * encoder tick count;
  *
  * @return
- * A double pointer that contains both encoder values.
+ * An integer pointer that contains both encoder values.
  */
-double* PeanutChassis::getEncoderValues()
+int* PeanutChassis::getEncoderValues()
 {
-	return new double[leftTalon->GetSelectedSensorPosition(0), rightTalon->GetSelectedSensorPosition(0)];
+	return new int[leftTalon->GetSelectedSensorPosition(0), rightTalon->GetSelectedSensorPosition(0)];
 }
 
 /**
